feat: useUserChallenge hook
